Suggest adding some descriptive text
I have put together some draft text introducing the dataset for this README file.  This way, hopefully it will be easier for people to find and to understand what they are accessing.  I drafted text using the forestgeo.si.edu site but please feel free to edit this as you see fit!  Thank you for putting together this resource!